refactor: moved getFormInitialValues to util
